ai_service/requirements.txt
fastapi==0.115.6
httpx==0.28.1
uvicorn[standard]==0.34.0
faiss-cpu>=1.7.4
numpy>=1.24.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
